testPermalinksListOnStatusPage() was moved to the FreestyleProjectTest class
FreestyleProject7Test was deleted
